Rename android.utils.Pools to com.android.photos.util.Pools
android.utils.Pools is a class copied from frameworks/base. It was
copied because the framework one has been marked as @hide. This caused
the class to be duplicated into the Gallery2 apk and the boot classpath.
At run-time, the class from bootclasspath was used because the
classloader tries to load a class from the parent class loader first.

However, this combined with the recent activation of the kill-switch
caused a runtime error. Working around the issue by renaming the class
copied into the app so that it does not collide with the
framework-defined one.

Bug: 77544391
Test: build aosp_walleye. Take a video and see it via the Gallery2.
The app does not crash.

Change-Id: Ia0af4da84242044da2e0ab63e38df243d9769418